pokemon searching using api project
This is a simple website created using HTML, CSS, and JS. This website is created for POKEMON fans.  The main motive of this site is to show pokemon data to pokemon fans by simply search pokemon's  names in the search bar. By using this website knowing about the different pokemons is easy and efficient.